requirements.txt
pause==0.3
PyYAML==6.0
pyzotero==1.5.5
requests==2.22.0
